requirements.txt
streamlit==1.26.0
requests==2.31.0
pdfplumber==0.9.0

# LangChain modular packages
langchain-core==0.2.43
langchain-community==0.2.19
langchain-text-splitters==0.2.4

sentence-transformers==2.2.2
faiss-cpu==1.13.2

google-generativeai==0.8.6
